feat: Add doctor command for health checks
Adds a new `doctor` command to `age-store.py` to run
a series of health checks on the user's environment and
configuration.

The `doctor` command checks for:
- Presence of `age` and `age-keygen` executables.
- Permissions of the unencrypted user secret file.
- Encryption status of the user secret.
- Ability to load the user's private key.
- Whether the current user is listed in `users.json`.
- Ability to decrypt the master key.

Also refactors the user secret permission check in
`read_user_secret` to use a new dedicated function,
which is reused in `cmd_doctor` function.