tests: Replace tests.py with shellspec tests
shellspec is a testing tool for CLI apps.
refer to shellspec-syntax.txt to learn how to define tests using
shellspec.
age-store-tests.txt contains shellspec tests definitions for age-store.

We also switched from scenario based tests in tests.py to more feature
oriented tests in age-store-tests.txt.
This is because it makes it easier to add new tests when introducing new
features. It also keeps each individual test size to be small and
manageable.